GDB-5944 Creates CLI for the initial functionalities for OntoRefine
- Added dependency to the [picocli](https://picocli.info/) library,
which will be used to create the actual CLI commands for the
communication with the Refine instance.
- Introduced few new commands and CLI processes for them. The commands
are:
  - export of the project operations. The operations are provided in
JSON format.
  - export of the project data in RDF format. The command uses the
project operations to extract the operation for the RDF mapping created
via RDF-mapping tool (integrated as a part of the Refine instance) and
transform the data.
  - reconciliation of the project data although it has only the initial
structure and it will be finished later.
- Added tests for the added CLI processes.
- Cleaned all issues reported by the checkstyle plugin, which allowed us
to enable it as a part of the building process in order to keep the code
base clean from now on.
- Added missing JavaDoc where it was required.