GDB-6097 Changes the repository where the package will be released
- Ceased the tries to create a package with the library in the GitHub
repository, because we found out that the upload of packages is disabled
for the organization (it might be related to our payment plan or more
likely the lack of it).
- Changed the repository where the package will be released to our
public Maven.
- Configured the new address for the distribution management in the pom.
- Configured new user and address in the release workflow.
- Reverted the CI workflow.
- Updated the release workflow to skip the tests and the checkstyle,
when packaging and deploying, because the CI pipeline guarantees that
the master is always in green state. Furthermore this will reduce the
time for the release process.