GDB-6140 Introduces command for reconciliation service registration
- Introduces few new commands, which aim to allow registration of
reconciliation service as standard so that it can be used in reconcile
operations. The command which does the registration is composition of
the commands for retrieving and setting specific preference property
called `reconciliation.standardServices`. Additionally the command will
try to retrieve the registration information about the service by
executing a `GET` request with the provided service `URI`.
- Bumped up the version of the library to `1.1.0` from `1.0.0` as this
is a new functionality that was not planned initially.
- Added unit tests for the commands in order to verify their behavior
and to guarantee that future updates will not break something in the
general cases.