GDB-6161 Fixes the command for operations exporting
- The issue with the command was that the returned JSON from the Refine
tool was containing additional tags and operation wrapping objects,
which causes failure if a client tries to use the JSON directly in the
apply operations command.

Apparently there is an additional processing of the JSON in the UI in
the export window, which is the reason the exported JSONs were
different.

To fix the issue we are processing the response from the Refine tool by
extracting all of the objects which are under `operation` tag and
collect them in a new array, which is then returned as result from the
operation.
- Added tests to verify the behavior of the command and the changes that
were added.