Adds support for setting additional configuration for ExportRowsCommand
- Moved the `ExportRowsCommand` in dedicated package in order to comply
with the structure of the project and the rest of the commands.
- Renamed the `UploadOptions` to only `Options`. The new interface is
intended to be used as general type for providing additional
configurations for all commands that are requiring in some way
additional configuring.
- Added property of type `Options` to the `ExportRowsCommand` which will
allow to pass additional parameters for the process like: `quoteAll`,
`separator`, etc.
- Added enum `Engines` which will contain the options for the export
engines. It is convenient to have such enumeration, because the engines
are basically static JSON objects that need to be passed as request
parameters.