Introduces few enhancements, couple of bug fixes and changes
- Enhanced the `RefineClients` type with additional methods, which are
producing different types of `RefineClient` instances. The intention
behind this change is to allow the developers to define their own way of
securing the client or pre-configure different aspects of it, like
timeouts, connections, etc.
- Fix `setOptions` method in `ExportRowsCommand#Builder` to follow the
fluent pattern.
- Reverted the wrongly committed changes to the `RefineCleint` type.
- Fixed broken integration tests for reconciliation registration due to
changes in the response from the reconciliation address. We are now
going to use our own services for this test as they are public and
usable.