Introduces changes in the RDF export result storing and retrieval
- Changed the storing and the retrieval mechanisms for the RDF export
result. Now the buffering of the result will depend on the length of the
content of the return HTTP entity and/or if the user explicitly requests
the result in specific type.

The options for the return type were reduced and move as command
property, instead of sub-property of the ``ResultFormat``. This makes
the configuration more transparent and simple to use.

Additionally, now the result of the export can be retrieved from all
available methods in the response, which is more continent in specific
cases.
Exception to that is the case, where the result is stored in a file and
the content exceeds the ``String`` buffer size. When such case occurs
and the user retrieves the result as ``String``, OOM will be thrown.